% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_1.R
\name{payment}
\alias{payment}
\title{Create a Boxplot}
\usage{
payment(payment_type)
}
\arguments{
\item{payment}{Type of payment - "medicare", "total", "covered"}
}
\value{
Boxplot
}
\description{
This function creates a boxplot for DRG data by taking input of type of payment
}
