% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_2.R
\name{stats}
\alias{stats}
\title{Function calculates either the mean, median, or standard deviation of the DRG
codes. over all of the DRG codes for average Medicare payments}
\usage{
stats(stat_type = "mean")
}
\arguments{
\item{statistics}{is a string which specifies the statistic and can take
values - mean, medium and standard deviation}
}
\value{
returns the chosen statistic calculated
}
\description{
Function calculates either the mean, median, or standard deviation of the DRG
codes. over all of the DRG codes for average Medicare payments
}
